Refactored compiler, many other small tweaks.
The compiler parts have now been united as the front- and backend
modules. Each module is only one source file. In addition the interface
to the compiler now only features two functions, one for each of the
two ends, freening the caller from dirty details.

Function derivation is now moved into the syntax node module since it
is useless in any other context.

The program no longer exits on error, instead error are reported and
control is reverted back to the caller of the failed function. This
would allow for proper exception handling.

Undefined symbols, nodes and so on are now caught and can be handled
gracefully as well.

Countless other minor cleanups.